requirements.txt
# Core dependencies
streamlit>=1.28.0
sqlite-utils>=3.34
pandas>=2.0.0
numpy>=1.24.0

# LLM and NLP
langchain>=0.1.0
llama-index>=0.9.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0

# PDF processing
pdfplumber>=0.7.6
pdfminer.six>=20221105
PyPDF2>=3.0.1

# File handling and utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
jsonschema>=4.17.0
email-validator>=2.1.0

# HTTP client for LLM APIs
requests>=2.28.0
httpx>=0.24.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced LLM integrations
# openai>=1.0.0  # Uncomment if using OpenAI
# anthropic>=0.7.0  # Uncomment if using Anthropic